DueDatePicker: display selection hints
This will display either a placeholder message, or the full date and
time of the set due date.
