+++ pootle/templates/layout.html
@@ -29,6 +29,7 @@
       MARKUP_FILTER: '{{ settings.POOTLE_MARKUP_FILTER }}',
       SOCIAL_AUTH_PROVIDERS: {{ SOCIAL_AUTH_PROVIDERS|to_js }},
       UI_LOCALE: '{{ LANGUAGE_CODE }}',
+      TIME_ZONE: '{{ settings.TIME_ZONE }}',
     };
     </script>
